Refactor: Move the custom endpoints in the action package.
