+++ alert-common/src/main/java/com/synopsys/integration/alert/common/action/endpoint/ButtonCustomEndpoint.java
@@ -20,31 +20,14 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.common.descriptor.config.field.endpoint.table;
+package com.synopsys.integration.alert.common.action.endpoint;
 
-import java.util.List;
-
-import org.springframework.http.HttpStatus;
-
-import com.synopsys.integration.alert.common.action.ActionResult;
 import com.synopsys.integration.alert.common.action.CustomEndpointManager;
-import com.synopsys.integration.alert.common.descriptor.config.field.endpoint.SimpleCustomEndpoint;
 import com.synopsys.integration.alert.common.exception.AlertException;
 
-public abstract class TableSelectCustomEndpoint extends SimpleCustomEndpoint<List<?>> {
-
-    protected TableSelectCustomEndpoint(String fieldKey, CustomEndpointManager customEndpointManager) throws AlertException {
+public abstract class ButtonCustomEndpoint extends CustomEndpoint<String> {
+    public ButtonCustomEndpoint(String fieldKey, CustomEndpointManager customEndpointManager) throws AlertException {
         super(fieldKey, customEndpointManager);
-    }
-
-    @Override
-    protected ActionResult<List<?>> createErrorResponse(Exception e) {
-        return new ActionResult<>(HttpStatus.INTERNAL_SERVER_ERROR, String.format("An internal issue occurred while trying to retrieve your select data: %s", e.getMessage()));
-    }
-
-    @Override
-    protected ActionResult<List<?>> createSuccessResponse(List<?> response) {
-        return new ActionResult<>(HttpStatus.OK, response);
     }
 
 }
