+++ alert-common/src/main/java/com/synopsys/integration/alert/common/action/endpoint/SelectCustomEndpoint.java
@@ -20,14 +20,32 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.common.descriptor.config.field.endpoint;
+package com.synopsys.integration.alert.common.action.endpoint;
 
+import java.util.List;
+
+import org.springframework.http.HttpStatus;
+
+import com.synopsys.integration.alert.common.action.ActionResult;
 import com.synopsys.integration.alert.common.action.CustomEndpointManager;
+import com.synopsys.integration.alert.common.descriptor.config.field.LabelValueSelectOption;
 import com.synopsys.integration.alert.common.exception.AlertException;
 
-public abstract class ButtonCustomEndpoint extends CustomEndpoint<String> {
-    public ButtonCustomEndpoint(String fieldKey, CustomEndpointManager customEndpointManager) throws AlertException {
+public abstract class SelectCustomEndpoint extends SimpleCustomEndpoint<List<LabelValueSelectOption>> {
+
+    public SelectCustomEndpoint(String fieldKey, CustomEndpointManager customEndpointManager) throws AlertException {
         super(fieldKey, customEndpointManager);
+    }
+
+    @Override
+    protected ActionResult<List<LabelValueSelectOption>> createErrorResponse(Exception e) {
+        return new ActionResult<>(HttpStatus.INTERNAL_SERVER_ERROR, String.format("An internal issue occurred while trying to retrieve your select data: %s", e.getMessage()));
+    }
+
+    @Override
+    protected ActionResult<List<LabelValueSelectOption>> createSuccessResponse(List<LabelValueSelectOption> response) {
+
+        return new ActionResult<>(HttpStatus.OK, response);
     }
 
 }
