+++ alert-common/src/main/java/com/synopsys/integration/alert/common/action/endpoint/SimpleCustomEndpoint.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.common.descriptor.config.field.endpoint;
+package com.synopsys.integration.alert.common.action.endpoint;
 
 import org.springframework.web.server.ResponseStatusException;
