+++ alert-common/src/main/java/com/synopsys/integration/alert/common/action/endpoint/table/TableSelectCustomEndpoint.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.common.descriptor.config.field.endpoint;
+package com.synopsys.integration.alert.common.action.endpoint.table;
 
 import java.util.List;
 
@@ -28,23 +28,22 @@ import org.springframework.http.HttpStatus;
 
 import com.synopsys.integration.alert.common.action.ActionResult;
 import com.synopsys.integration.alert.common.action.CustomEndpointManager;
-import com.synopsys.integration.alert.common.descriptor.config.field.LabelValueSelectOption;
+import com.synopsys.integration.alert.common.action.endpoint.SimpleCustomEndpoint;
 import com.synopsys.integration.alert.common.exception.AlertException;
 
-public abstract class SelectCustomEndpoint extends SimpleCustomEndpoint<List<LabelValueSelectOption>> {
+public abstract class TableSelectCustomEndpoint extends SimpleCustomEndpoint<List<?>> {
 
-    public SelectCustomEndpoint(String fieldKey, CustomEndpointManager customEndpointManager) throws AlertException {
+    protected TableSelectCustomEndpoint(String fieldKey, CustomEndpointManager customEndpointManager) throws AlertException {
         super(fieldKey, customEndpointManager);
     }
 
     @Override
-    protected ActionResult<List<LabelValueSelectOption>> createErrorResponse(Exception e) {
+    protected ActionResult<List<?>> createErrorResponse(Exception e) {
         return new ActionResult<>(HttpStatus.INTERNAL_SERVER_ERROR, String.format("An internal issue occurred while trying to retrieve your select data: %s", e.getMessage()));
     }
 
     @Override
-    protected ActionResult<List<LabelValueSelectOption>> createSuccessResponse(List<LabelValueSelectOption> response) {
-
+    protected ActionResult<List<?>> createSuccessResponse(List<?> response) {
         return new ActionResult<>(HttpStatus.OK, response);
     }
